+++ app/src/main/java/com/example/swith/data/ProfileRequest.kt
@@ -1,5 +1,5 @@
 package com.example.swith.data
 
 data class ProfileRequest(
-    val userIdx: Int
+    val userIdx: Long
 )
